Heisenberg_Protocol/backend/requirements.txt
PyJWT
SQLAlchemy
fastapi
greenlet
pydantic
python-dotenv
python-jose[cryptography]
sqlalchemy
uvicorn